\name{BayesDLMfMRI-package}
\alias{BayesDLMfMRI-package}
\alias{BayesDLMfMRI}
\docType{package}
\title{
\packageTitle{BayesDLMfMRI}
}
\description{
\packageDescription{BayesDLMfMRI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesDLMfMRI}
\packageIndices{BayesDLMfMRI}

}
\author{
\packageAuthor{BayesDLMfMRI}

Maintainer: \packageMaintainer{BayesDLMfMRI}
}
\references{
Package source is in the help file for package.skeleton
}

\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
 #simple examples of the most important functions 
}
