% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceNeighbors.R
\name{.distanceNeighbors}
\alias{.distanceNeighbors}
\title{.distanceNeighbors}
\usage{
.distanceNeighbors(posi.refer, r1)
}
\arguments{
\item{posi.refer}{the position of the voxel in the brain image.}

\item{r1}{a positive integer number that defines the distance from every voxel with its most distant neighbor. This value determines the size of the cluster. The users can set a range of different r values: r = 0, 1, 2, 3, 4, which leads to q = 1, 7, 19, 27, 33, where q is the size of the cluster.}
}
\description{
this is an internal function
}
\details{
this is an internal function
}
\references{
\insertRef{CARDONAJIMENEZ2021107297}{BayesDLMfMRI}

\insertRef{cardona2021bayesdlmfmri}{BayesDLMfMRI}
}
\keyword{internal}
