% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\docType{data}
\name{ffd}
\alias{ffd}
\title{ffd
example overlay used in the vignettes.}
\format{
An object of class \code{nifti} of dimension 91 x 109 x 91.
}
\usage{
ffd
}
\description{
ffd
example overlay used in the vignettes.
}
\examples{
data("ffd", package="BayesDLMfMRI")
}
\keyword{datasets}
