% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_example_fMRI_data}
\alias{get_example_fMRI_data}
\title{get_example_fMRI_data}
\usage{
get_example_fMRI_data(save_path = "./example_data", force = FALSE)
}
\arguments{
\item{save_path}{location where the data the example data is stored.}

\item{force}{force the download, even if the data already exists.}
}
\description{
This function is used to download the example data used in the Vignettes.
}
\details{
By default this function
}
