% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\docType{data}
\name{mask}
\alias{mask}
\title{A 3D array that works as a brain of reference (MNI atlas).}
\format{
An object of class \code{nifti} of dimension 91 x 109 x 91.
}
\usage{
mask
}
\description{
A 3D array that works as a brain of reference (MNI atlas) for the group analysis.
}
\examples{
data("mask", package="BayesDLMfMRI")
}
\keyword{datasets}
