% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\name{print.fMRI_group_single_voxel}
\alias{print.fMRI_group_single_voxel}
\title{print.fMRI_group_single_voxel}
\usage{
\method{print}{fMRI_group_single_voxel}(x, ...)
}
\arguments{
\item{x}{is the returned value of any of the GroupSingleVoxel* functions}

\item{...}{Other potential arguments}
}
\description{
Print the structure of the object related to the SingleVoxel* functions.
}
\details{
Print the structure of the object related to the GroupSingleVoxel* functions.
}
\examples{
\dontrun{
DatabaseGroup <- get_example_fMRI_data_group()
data("covariates", package="BayesDLMfMRI")
res <- GroupSingleVoxelFFBS(posi.ffd = c(14, 56, 40), DatabaseGroup,
                            covariates = Covariates, m0 = 0, Cova = 100, 
                            delta = 0.95, S0 = 1, n0 = 1, N1 = FALSE, 
                            Nsimu1 = 100, r1 = 1, Cutpos = 30)
print(res)
}
}
