% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\name{print.fMRI_single_evidence}
\alias{print.fMRI_single_evidence}
\title{print.fMRI_single_evidence}
\usage{
\method{print}{fMRI_single_evidence}(x, ...)
}
\arguments{
\item{x}{is the returned value of any of the ffdEvidence* functions}

\item{...}{Other potential arguments}
}
\description{
Print the structure of the object related to the ffdEvidence* functions.
}
\details{
Print the structure of the object related to the ffdEvidence* functions.
}
\examples{
\dontrun{
fMRI.data  <- get_example_fMRI_data()
data("covariates", package="BayesDLMfMRI")
res <- ffdEvidenceFFBS(ffdc = fMRI.data, covariates = Covariates, m0=0, Cova=100,
                       delta=0.95, S0=1, n0=1, N1=FALSE, 
                       Nsimu1 = 100, Cutpos1=30, r1 = 1,
                       perVol = 0.10, Ncores=3)
print(res)
}
}
