% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\name{print.fMRI_single_voxel}
\alias{print.fMRI_single_voxel}
\title{print.fMRI_single_voxel}
\usage{
\method{print}{fMRI_single_voxel}(x, ...)
}
\arguments{
\item{x}{is the returned value of any of the SingleVoxel* functions,}

\item{...}{Other potential arguments}
}
\description{
Print the structure of the object related to the SingleVoxel* functions.
}
\details{
Print the structure of the object related to the SingleVoxel* functions.
}
\examples{
\dontrun{
fMRI.data  <- get_example_fMRI_data()
data("covariates", package="BayesDLMfMRI")
res.indi <- SingleVoxelFSTS(posi.ffd = c(14, 56, 40), 
                            covariates = Covariates,
                            ffdc =  fMRI.data, 
                            m0 = 0, Cova = 100, delta = 0.95, S0 = 1, 
                            n0 = 1, Nsimu1 = 100, N1 = N1, Cutpos1 = 30, 
                            Min.vol = 0.10, r1 = 1)
print(res.indi)
}
}
