% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\name{summary.fMRI_group_evidence}
\alias{summary.fMRI_group_evidence}
\title{summary.fMRI_group_evidence}
\usage{
\method{summary}{fMRI_group_evidence}(object, ...)
}
\arguments{
\item{object}{is the returned value of any of the fdGroupEvidence* functions}

\item{...}{Other potential arguments}
}
\description{
Summary function
}
\details{
Summary function
}
\examples{
\dontrun{
DatabaseGroup <- get_example_fMRI_data_group()
data("covariates", package="BayesDLMfMRI")
data("mask", package="BayesDLMfMRI")
res <- ffdGroupEvidenceFETS(ffdGroup = DatabaseGroup, covariates = Covariates, 
                            m0 = 0, Cova = 100, delta = 0.95, S0 = 1, 
                            n0 = 1, N1 = FALSE, Nsimu1 = 100, Cutpos=30, 
                            r1 = 1, Test = "JointTest", mask = mask, Ncores = 7)
summary(res)
}
}
