% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\name{summary.fMRI_single_evidence}
\alias{summary.fMRI_single_evidence}
\title{summary.fMRI_single_evidence}
\usage{
\method{summary}{fMRI_single_evidence}(object, ...)
}
\arguments{
\item{object}{is the returned value of any of the ffdEvidence* functions}

\item{...}{Other potential arguments}
}
\description{
Summary function
}
\details{
Summary function
}
\examples{
\dontrun{
fMRI.data  <- get_example_fMRI_data()
data("covariates", package="BayesDLMfMRI")
res <- ffdEvidenceFFBS(ffdc = fMRI.data, covariates = Covariates, m0=0, Cova=100,
                       delta=0.95, S0=1, n0=1, N1=FALSE, 
                       Nsimu1 = 100, Cutpos1=30, r1 = 1,
                       perVol = 0.10, Ncores=3)
summary(res)
}
}
