\name{BayesDLMfMRI-package}
\alias{BayesDLMfMRI-package}
\alias{BayesDLMfMRI}
\docType{package}
\title{
\packageTitle{BayesDLMfMRI}
}
\description{
\packageDescription{BayesDLMfMRI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesDLMfMRI}
\packageIndices{BayesDLMfMRI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesDLMfMRI}

Maintainer: \packageMaintainer{BayesDLMfMRI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
