% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\docType{data}
\name{Covariates}
\alias{Covariates}
\title{Covariates related to the observed BOLD response}
\format{
An object of class \code{data.frame} with 310 rows and 2 columns.
}
\usage{
Covariates
}
\description{
Covariates related to the observed BOLD response and its derivative used in the examples presented in the vignettes.
}
\examples{
data("covariates", package="BayesDLMfMRI")
}
\keyword{datasets}
