% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDLMfMRI.R
\docType{data}
\name{ffd}
\alias{ffd}
\title{MNI image used to plot posterior probability maps in the vignette examples.}
\format{
An object of class \code{nifti} of dimension 91 x 109 x 91.
}
\usage{
ffd
}
\description{
MNI image used to plot posterior probability maps in the examples presented in the vignettes.
}
\examples{
data("ffd", package="BayesDLMfMRI")
}
\keyword{datasets}
